+++ bindings/java/plugin/src/it/workspace-build-tag-env-set/.workspace.json
