+++ bindings/java/plugin/src/it/workspace-build-tag-env-set/invoker.properties
@@ -1,0 +1,3 @@
+invoker.goals = -B -V package
+invoker.buildResult = success
+invoker.environmentVariables.BUILD_TAG = 99
