+++ bindings/java/plugin/src/it/workspace-build-tag-env-set/project1/.project.json
