+++ bindings/java/plugin/src/it/workspace-build-tag-env-unset/invoker.properties
@@ -1,0 +1,2 @@
+invoker.goals = -B -V package
+invoker.buildResult = success
