+++ bindings/java/plugin/src/it/workspace-build-tag-env-unset/project1/.project.json
@@ -1,0 +1,5 @@
+{
+  "name": "project1",
+  "version": "1.0.0",
+  "usage": []
+}
