+++ bindings/java/plugin/src/it/workspace-build-tag-env-unset/project1/a.sysml
@@ -1,0 +1,1 @@
+package A;
