--- bindings/java/plugin/src/it/workspace-build-tag-env/invoker.properties
@@ -1,10 +1,0 @@
-# Invocation 1: BUILD_TAG is set — the with-build-tag profile activates and the
-# built KPAR version receives the dev.99 suffix.
-invoker.1.goals = -B -V package
-invoker.1.buildResult = success
-invoker.environmentVariables.BUILD_TAG.1=99
-
-# Invocation 2: BUILD_TAG is absent — the with-build-tag profile is inactive and
-# the built KPAR version is unchanged.
-invoker.2.goals = -B -V package
-invoker.2.buildResult = success
